Added Viewer::getEvents() and moved the PrintVisitor from the Viewer::pollEvents() into it's own include/vsg/ui/PrintEvents.h header
